+++ R/trendLevel.R
@@ -135,12 +135,13 @@ trendLevel <- function(
   n.levels = c(10, 10, 4),
   windflow = NULL,
   limits = NULL,
+  breaks = NULL,
+  trans = FALSE,
   min.bin = 1,
   cols = "default",
   auto.text = TRUE,
   key.title = paste("use.stat.name", pollutant, sep = " "),
   key.position = "right",
-  breaks = NULL,
   statistic = c(
     "mean",
     "max",
@@ -579,7 +580,8 @@ trendLevel <- function(
         colours = resolve_colour_opts(cols, 100),
         na.value = col.na,
         oob = scales::oob_squish,
-        limit = limits
+        limit = limits,
+        transform = get_scale_transform(trans)
       ) +
       ggplot2::guides(
         fill = ggplot2::guide_colorbar(